refactor(Squib Games): Moved logic into utils files for clarity & improved development
